requirements.txt
alabaster==0.7.8
Babel==2.3.4
dj-database-url==0.4.1
Django==1.9.7
djangorestframework==3.3.3
docutils==0.12
gunicorn==19.6.0
imagesize==0.7.1
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
Pygments==2.1.3
pytz==2016.4
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.3
whitenoise==3.2
